refactor: 관리자: 강좌 상세페이지, 강좌 개설페이지 / 사용자: 강좌 개설페이지 리펙토링
